Add files via upload
This is a *first of kind as far as I can tell* file creator for interfacing with the BRENDA Enzyme Database. 

It takes the entire JSON file, available for download on BRENDA's website (https://www.brenda-enzymes.org/download.php) and extracts just the inhibitors and the Enzymes (EC's) they inhibit. 

Can be tweaked for anything on the same JSON key level as 'inhibitor'.